requirement.txt
asn1crypto==0.24.0
bcrypt==3.1.4
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.3.1
Django==2.1
django-appconf==1.0.2
django-compressor==2.2
django-filter==2.0.0
django-rest-swagger==2.2.0
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
Fabric==1.14.0
Fabric3==1.14.post1
idna==2.7
invoke==1.1.1
ipython-genutils==0.2.0
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.0
munch==2.3.2
mysqlclient==1.3.13
openapi-codec==1.3.2
paramiko==2.4.1
pyasn1==0.4.4
pycparser==2.18
PyJWT==1.6.4
PyNaCl==1.2.1
pytz==2018.5
rcssmin==1.0.6
requests==2.19.1
rjsmin==1.0.12
rsa==3.4.2
shortuuid==0.5.0
simplejson==3.16.0
six==1.11.0
uritemplate==3.0.0
urllib3==1.23
